requirements/base.txt
Django==2.1.1
psycopg2-binary==2.7.5
dj-database-url==0.5.0
whitenoise==4.0
django-allauth==0.37.1
django-extensions==2.1.2
werkzeug==0.14.1
pyopenssl==18.0.0
django-js-reverse==0.8.2
djangorestframework==3.8.2
django-colorfield==0.1.15
raven==6.9.0
git+https://github.com/SalesforceFoundation/Cumulusci.git@feature/python3
GitPython==2.1.11
django-rq==1.2.0
